various refactors for new package name
