+++ support/cas-server-support-gua/src/main/java/org/apereo/cas/impl/LdapUserGraphicalAuthenticationRepository.java
@@ -1,7 +1,7 @@
-package org.apereo.cas.gua.impl;
+package org.apereo.cas.impl;
 
 import org.apereo.cas.configuration.CasConfigurationProperties;
-import org.apereo.cas.gua.api.UserGraphicalAuthenticationRepository;
+import org.apereo.cas.api.UserGraphicalAuthenticationRepository;
 import org.apereo.cas.util.CollectionUtils;
 import org.apereo.cas.util.LdapUtils;
